services/bid-worker/requirements.txt
pydantic==2.10.3
pydantic-settings==2.6.1
confluent-kafka==2.6.1
redis==5.2.1
psycopg2-binary==2.9.10
